Add files via upload
As an Analyst, I need to develop a data analysis tool for analyzing the overall situation of orders and their conclusion for various courier agencies in South Africa and Botswana.  Microsoft Power BI was used to analyze and visualize the orders across states within countries, highlight the top performing clients, the origin of late orders and the yearly orders.